GameManager, CSV 무게, 가방 float화

+++ Assets/3.Script/B/1_GameManager/GameManager.cs
@@ -10,9 +10,9 @@ public class GameManager : MonoBehaviour
     public int P_Str = 1; //힘
     public int P_Spd = 5; //속도
     public int P_Money = 1000; //소지 돈
-    public int P_Maxbag = 100; //가방최대무게
+    public float P_Maxbag = 100; //가방최대무게
 
-    public int P_Weight = 0; //현재 무게
+    public float P_Weight = 0; //현재 무게
 
     public bool P_isEnglish; //한 영문전환
